Initial import of the pagure-taiga 3rd party plugin
This plugin is used to integrate pagure with a taiga instance.
It does a 2 way sync from pagure to taiga and from taiga to pagure. It
supports syncing tickets between both apps as well as comments on these
tickets.
In addition, when a ticket changes status in addition, these status
changes are reflected as tags in the pagure UI. Similarly, if a ticket
is tagged with one of the status used in taiga, it will be updated to
that status in taiga.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>